requirement.txt
fastapi==0.135.1
uvicorn==0.41.0
SQLAlchemy==2.0.48
PyMySQL==1.1.2
passlib==1.7.4
bcrypt==4.3.0
python-jose==3.5.0
pydantic==2.12.5
email-validator==2.3.0
python-multipart==0.0.22